Split llava from yolo2
