+++ .env
@@ -1,0 +1,1 @@
+SCREENSHOT_CAM_FILE=img/screen_cam.jpg
