+++ requirements.txt
@@ -3,3 +3,4 @@ opencv-python
 numpy
 ollama
 pyserial
+dotenv
